Make the class Film use 0-1 colors instead of 0-255 ones

+++ proj01/src/scene/film.hpp
@@ -18,7 +18,7 @@ class Film{
 public:
 	int width;
 	int height;
-	int*** buff;
+	float*** buff;
 	FilmType type;
 	string filename;
 	string imgType;
